+++ client/src/tls/mod.rs
@@ -20,6 +20,6 @@ mod tls_client_connection;
 mod tls_client_stream;
 mod tls_stream;
 
-pub use self::tls_client_connection::TlsClientConnection;
-pub use self::tls_client_stream::TlsClientStream;
-pub use self::tls_stream::TlsStream;
+pub use self::tls_client_connection::{TlsClientConnection, TlsClientConnectionBuilder};
+pub use self::tls_client_stream::{TlsClientStream, TlsClientStreamBuilder};
+pub use self::tls_stream::{TlsStream, TlsStreamBuilder};