+++ client/src/tls/tls_client_connection.rs
@@ -19,13 +19,16 @@ use std::io;
 
 use futures::Future;
 use native_tls::Pkcs12;
+#[cfg(target_os = "linux")]
+use openssl::x509::X509 as OpensslX509;
+#[cfg(target_os = "macos")]
 use security_framework::certificate::SecCertificate;
 use tokio_core::net::TcpStream;
 use tokio_core::reactor::Core;
 
 use ::error::*;
 use ::client::{ClientConnection, ClientStreamHandle};
-use ::tls::TlsClientStream;
+use ::tls::{TlsClientStream, TlsClientStreamBuilder};
 
 /// TCP based DNS client
 pub struct TlsClientConnection {
@@ -35,22 +38,8 @@ pub struct TlsClientConnection {
 }
 
 impl TlsClientConnection {
-  /// Creates a new client connection.
-  ///
-  /// *Note* this has side affects of establishing the connection to the specified DNS server and
-  ///        starting the event_loop. Expect this to change in the future.
-  ///
-  /// # Arguments
-  ///
-  /// * `name_server` - address of the name server to use for queries
-  pub fn new(name_server: SocketAddr,
-             subject_name: String,
-             certs: Vec<SecCertificate>,
-             pkcs12: Option<Pkcs12>) -> ClientResult<Self> {
-    let io_loop = try!(Core::new());
-    let (tls_client_stream, handle) = TlsClientStream::new_tls(name_server, subject_name, io_loop.handle(), certs, pkcs12);
-
-    Ok(TlsClientConnection{ io_loop: io_loop, tls_client_stream: tls_client_stream, client_stream_handle: handle })
+  pub fn builder() -> TlsClientConnectionBuilder {
+    TlsClientConnectionBuilder(TlsClientStream::builder())
   }
 }
 
@@ -59,5 +48,39 @@ impl ClientConnection for TlsClientConnection {
 
   fn unwrap(self) -> (Core, Box<Future<Item=Self::MessageStream, Error=io::Error>>, Box<ClientStreamHandle>) {
     (self.io_loop, self.tls_client_stream, self.client_stream_handle)
+  }
+}
+
+pub struct TlsClientConnectionBuilder(TlsClientStreamBuilder);
+
+impl TlsClientConnectionBuilder {
+  #[cfg(target_os = "macos")]
+  pub fn add_ca(&mut self, ca: SecCertificate) {
+    self.0.add_ca(ca);
+  }
+
+  #[cfg(target_os = "linux")]
+  pub fn add_ca(&mut self, ca: OpensslX509) {
+    self.0.add_ca(ca);
+  }
+
+  /// Client side identity for client auth in TLS (aka mutual TLS auth)
+  pub fn identity(&mut self, pkcs12: Pkcs12) {
+    self.0.identity(pkcs12);
+  }
+
+  /// Creates a new client connection.
+  ///
+  /// *Note* this has side affects of establishing the connection to the specified DNS server and
+  ///        starting the event_loop. Expect this to change in the future.
+  ///
+  /// # Arguments
+  ///
+  /// * `name_server` - address of the name server to use for queries
+  pub fn build(self, name_server: SocketAddr, subject_name: String) -> ClientResult<TlsClientConnection> {
+    let io_loop = try!(Core::new());
+    let (tls_client_stream, handle) = self.0.build(name_server, subject_name, io_loop.handle());
+
+    Ok(TlsClientConnection{ io_loop: io_loop, tls_client_stream: tls_client_stream, client_stream_handle: handle })
   }
 }