+++ client/src/tls/tls_client_stream.rs
@@ -10,6 +10,9 @@ use std::io;
 
 use futures::{Async, Future, Poll, Stream};
 use native_tls::Pkcs12;
+#[cfg(target_os = "linux")]
+use openssl::x509::X509 as OpensslX509;
+#[cfg(target_os = "macos")]
 use security_framework::certificate::SecCertificate;
 use tokio_core::io::Io;
 use tokio_core::net::TcpStream as TokioTcpStream;
@@ -18,21 +21,38 @@ use tokio_tls::TlsStream as TokioTlsStream;
 
 use ::BufClientStreamHandle;
 use ::tcp::TcpClientStream;
-use ::tls::TlsStream;
+use ::tls::{TlsStream, TlsStreamBuilder};
 use ::client::ClientStreamHandle;
 
 pub type TlsClientStream = TcpClientStream<TokioTlsStream<TokioTcpStream>>;
 
 impl TlsClientStream {
-  /// it is expected that the resolver wrapper will be responsible for creating and managing
-  ///  new TcpClients such that each new client would have a random port (reduce chance of cache
-  ///  poisoning)
-  pub fn new_tls(name_server: SocketAddr,
-                 subject_name: String,
-                 loop_handle: Handle,
-                 certs: Vec<SecCertificate>,
-                 pkcs12: Option<Pkcs12>) -> (Box<Future<Item=TlsClientStream, Error=io::Error>>, Box<ClientStreamHandle>) {
-    let (stream_future, sender) = TlsStream::new_tls(name_server, subject_name, loop_handle, certs, pkcs12);
+  pub fn builder() -> TlsClientStreamBuilder {
+    TlsClientStreamBuilder(TlsStream::builder())
+  }
+}
+
+pub struct TlsClientStreamBuilder(TlsStreamBuilder);
+
+impl TlsClientStreamBuilder {
+  #[cfg(target_os = "macos")]
+  pub fn add_ca(&mut self, ca: SecCertificate) {
+    self.0.add_ca(ca);
+  }
+
+  #[cfg(target_os = "linux")]
+  pub fn add_ca(&mut self, ca: OpensslX509) {
+    self.0.add_ca(ca);
+  }
+
+  /// Client side identity for client auth in TLS (aka mutual TLS auth)
+  pub fn identity(&mut self, pkcs12: Pkcs12) {
+    self.0.identity(pkcs12);
+  }
+
+  pub fn build(self, name_server: SocketAddr, subject_name: String, loop_handle: Handle)
+         -> (Box<Future<Item=TlsClientStream, Error=io::Error>>, Box<ClientStreamHandle>) {
+    let (stream_future, sender) = self.0.build(name_server, subject_name, loop_handle);
 
     let new_future: Box<Future<Item=TlsClientStream, Error=io::Error>> =
       Box::new(stream_future.map(move |tls_stream| {