+++ client/src/tls/tls_stream.rs
@@ -13,8 +13,13 @@ use futures::sync::mpsc::unbounded;
 use native_tls;
 use native_tls::TlsConnector;
 use native_tls::Pkcs12;
+#[cfg(target_os = "macos")]
 use native_tls::backend::security_framework::TlsConnectorBuilderExt;
+#[cfg(target_os = "linux")]
+use native_tls::backend::openssl::TlsConnectorBuilderExt;
 use native_tls::Protocol::Tlsv12;
+#[cfg(target_os = "linux")]
+use openssl::x509::X509 as OpensslX509;
 #[cfg(target_os = "macos")]
 use security_framework::certificate::SecCertificate;
 use tokio_core::net::TcpStream as TokioTcpStream;
@@ -27,6 +32,69 @@ use ::tcp::TcpStream;
 pub type TlsStream = TcpStream<TokioTlsStream<TokioTcpStream>>;
 
 impl TlsStream {
+  pub fn builder() -> TlsStreamBuilder {
+    TlsStreamBuilder { ca_chain: vec![], identity: None }
+  }
+
+  #[cfg(target_os = "linux")]
+  fn build(certs: Vec<OpensslX509>, pkcs12: Option<Pkcs12>) -> io::Result<TlsConnector> {
+    let mut builder = try!(TlsConnector::builder().map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e))));
+    try!(builder.supported_protocols(&[Tlsv12]).map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e))));
+
+    for cert in certs {
+      try!(builder.builder_mut().builder_mut().cert_store_mut().add_cert(cert).map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e))));
+    }
+
+    if let Some(pkcs12) = pkcs12 { try!(builder.identity(pkcs12).map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e)))); }
+    builder.build().map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e)))
+  }
+
+  #[cfg(target_os = "macos")]
+  fn build(certs: Vec<SecCertificate>, pkcs12: Option<Pkcs12>) -> io::Result<TlsConnector> {
+    let mut builder = try!(TlsConnector::builder().map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e))));
+    try!(builder.supported_protocols(&[Tlsv12]).map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e))));
+    builder.anchor_certificates(&certs);
+    if let Some(pkcs12) = pkcs12 { try!(builder.identity(pkcs12).map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e)))); }
+    builder.build().map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e)))
+  }
+
+  /// Initializes a TcpStream with an existing tokio_core::net::TcpStream.
+  ///
+  /// This is intended for use with a TcpListener and Incoming.
+  pub fn from_tls_stream(stream: TokioTlsStream<TokioTcpStream>, peer_addr: SocketAddr) -> (Self, BufStreamHandle) {
+    let (message_sender, outbound_messages) = unbounded();
+
+    let stream = TcpStream::from_stream_with_receiver(stream, peer_addr, outbound_messages);
+
+    (stream, message_sender)
+  }
+}
+
+pub struct TlsStreamBuilder {
+  #[cfg(target_os = "macos")]
+  ca_chain: Vec<SecCertificate>,
+
+  #[cfg(target_os = "linux")]
+  ca_chain: Vec<OpensslX509>,
+  identity: Option<Pkcs12>,
+}
+
+impl TlsStreamBuilder {
+  #[cfg(target_os = "macos")]
+  pub fn add_ca(&mut self, ca: SecCertificate) {
+    self.ca_chain.push(ca);
+  }
+
+  #[cfg(target_os = "linux")]
+  pub fn add_ca(&mut self, ca: OpensslX509) {
+    self.ca_chain.push(ca);
+  }
+
+  /// Client side identity for client auth in TLS (aka mutual TLS auth)
+  pub fn identity(&mut self, pkcs12: Pkcs12) {
+    self.identity = Some(pkcs12);
+  }
+
   /// Creates a new TlsStream to the specified name_server
   ///
   /// [RFC 7858](https://tools.ietf.org/html/rfc7858), DNS over TLS, May 2016
@@ -53,16 +121,15 @@ impl TlsStream {
   /// * `name_server` - IP and Port for the remote DNS resolver
   /// * `subject_name` - The Subject Public Key Info (SPKI) name as associated to a certificate
   /// * `loop_handle` - The reactor Core handle
-  /// * `certs` - list of trusted certificates authorities
+  /// * `certs` - list of trusted certificate authorities
   /// * `pkcs12` - optional client identity for client auth (i.e. for mutual TLS authentication)
   /// TODO: make a builder for the certifiates...
-  pub fn new_tls(name_server: SocketAddr,
-                 subject_name: String,
-                 loop_handle: Handle,
-                 certs: Vec<SecCertificate>,
-                 pkcs12: Option<Pkcs12>) -> (Box<Future<Item=TlsStream, Error=io::Error>>, BufStreamHandle) {
+  pub fn build(self,
+               name_server: SocketAddr,
+               subject_name: String,
+               loop_handle: Handle) -> (Box<Future<Item=TlsStream, Error=io::Error>>, BufStreamHandle) {
     let (message_sender, outbound_messages) = unbounded();
-    let tls_connector = match Self::build(certs, pkcs12) {
+    let tls_connector = match TlsStream::build(self.ca_chain, self.identity) {
       Ok(c) => c,
       Err(e) => return (Box::new(future::err(e).into_future().map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e)))),
                         message_sender)
@@ -81,26 +148,6 @@ impl TlsStream {
       })
       .map_err(|e| io::Error::new(io::ErrorKind::ConnectionRefused, format!("tls error: {}", e)))
     );
-
-    (stream, message_sender)
-  }
-
-  #[deprecated = "is this used?"]
-  fn build(certs: Vec<SecCertificate>, pkcs12: Option<Pkcs12>) -> native_tls::Result<TlsConnector> {
-    let mut builder = try!(TlsConnector::builder());
-    try!(builder.supported_protocols(&[Tlsv12]));
-    builder.anchor_certificates(&certs);
-    if let Some(pkcs12) = pkcs12 { try!(builder.identity(pkcs12)); }
-    builder.build()
-  }
-
-  /// Initializes a TcpStream with an existing tokio_core::net::TcpStream.
-  ///
-  /// This is intended for use with a TcpListener and Incoming.
-  pub fn from_tls_stream(stream: TokioTlsStream<TokioTcpStream>, peer_addr: SocketAddr) -> (Self, BufStreamHandle) {
-    let (message_sender, outbound_messages) = unbounded();
-
-    let stream = TcpStream::from_stream_with_receiver(stream, peer_addr, outbound_messages);
 
     (stream, message_sender)
   }
@@ -221,8 +268,16 @@ fn tls_client_stream_test(server_addr: IpAddr) {
   // the tests should run within 5 seconds... right?
   // TODO: add timeout here, so that test never hangs...
   // let timeout = Timeout::new(Duration::from_secs(5), &io_loop.handle());
+
+  #[cfg(target_os = "macos")]
   let trust_chain = SecCertificate::from_der(&cert_der).unwrap();
-  let (stream, sender) = TlsStream::new_tls(server_addr, subject_name.to_string(), io_loop.handle(), vec![trust_chain], None);
+
+  #[cfg(target_os = "linux")]
+  let trust_chain = OpensslX509::from_der(&cert_der).unwrap();
+
+  let mut builder = TlsStream::builder();
+  builder.add_ca(trust_chain);
+  let (stream, sender) = builder.build(server_addr, subject_name.to_string(), io_loop.handle());
 
   let mut stream = io_loop.run(stream).ok().expect("run failed to get stream");